+++ CHANGELOG.md
@@ -1,0 +1,18 @@
+# Changelog
+
+# v1.0.0
+
+Forked from [link](https://github.com/nanobox-io/golang-scribble)
+
+**Enhancements**
+
+* Added `go.mod` nad `go.sum` files
+* Updated to `go 1.19`
+* Switched logger to `logrus`
+* Split code into sub files.
+* Added `CHANGELOG.md` file
+* Added `Makefile` file.
+
+**Other**
+* Updated LICENSE
+* Updated README.md
