+++ go.mod
@@ -1,0 +1,7 @@
+module github.com/rocketblend/scribble
+
+go 1.19
+
+require github.com/sirupsen/logrus v1.9.0
+
+require golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
