+++ read.go
@@ -1,0 +1,73 @@
+package scribble
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"path/filepath"
+)
+
+// Read a record from the database
+func (d *Driver) Read(collection, resource string, v interface{}) error {
+	// ensure there is a place to save record
+	if collection == "" {
+		return fmt.Errorf("missing collection - no place to save record")
+	}
+
+	// ensure there is a resource (name) to save record as
+	if resource == "" {
+		return fmt.Errorf("missing resource - unable to save record (no name)")
+	}
+
+	// create full path to record and check to see if file exists
+	record := filepath.Join(d.dir, collection, resource)
+	if _, err := stat(record); err != nil {
+		return err
+	}
+
+	// read record from database
+	b, err := ioutil.ReadFile(record + ".json")
+	if err != nil {
+		return err
+	}
+
+	// unmarshal data
+	return json.Unmarshal(b, &v)
+}
+
+// ReadAll records from a collection; this is returned as a slice of strings because
+// there is no way of knowing what type the record is.
+func (d *Driver) ReadAll(collection string) ([]string, error) {
+	// ensure there is a collection to read
+	if collection == "" {
+		return nil, fmt.Errorf("missing collection - unable to record location")
+	}
+
+	// create full path to collection and check to see if directory exists
+	dir := filepath.Join(d.dir, collection)
+	if _, err := stat(dir); err != nil {
+		return nil, err
+	}
+
+	// read all the files in the transaction.Collection; an error here just means
+	// the collection is either empty or doesn't exist
+	files, _ := ioutil.ReadDir(dir)
+
+	// the files read from the database
+	var records []string
+
+	// iterate over each of the files, attempting to read the file. If successful
+	// append the files to the collection of read files
+	for _, file := range files {
+		b, err := ioutil.ReadFile(filepath.Join(dir, file.Name()))
+		if err != nil {
+			return nil, err
+		}
+
+		// append read file
+		records = append(records, string(b))
+	}
+
+	// unmarhsal the read files as a comma delimeted byte array
+	return records, nil
+}
